+++ hpx/runtime/parcelset/decode_parcels.hpp
@@ -221,7 +221,8 @@ namespace hpx { namespace parcelset
                                 "schedule_parcel",
                                 threads::pending, true,
                                 threads::thread_priority_boost,
-                                threads::thread_schedule_hint(num_thread),
+                                threads::thread_schedule_hint(
+                                    static_cast<std::int16_t>(num_thread)),
                                 threads::thread_stacksize_default);
                         }
                         // If we are the first deferred parcel, we don't need to spin
